+++ router/core/src/main/scala/io/buoyant/router/Router.scala
@@ -360,6 +360,7 @@ object StackRouter {
     stk.push(StatsFilter.module)
     stk.push(DstTracing.Path.module)
     stk.push(DstPathCtx.Setter.module)
+    stk.push(PathRegistry.module)
     stk.result
   }